Sketch out a different approach to a dotted key -based accessor.
Uses a generic function to completely separate the implementation for
different data types (dict and list for now) as well as providing a
"plugin" point for other types.

Provide a high-level set/get/whatever, function-based API as well as a
way to wrap an existing object to make it dotted-accessible.

Right now I have no idea if this is even useful. It's certainly not
doing everything a dottedish-user may want (but should be very easy to
extend) and it has no tests.