Allow the creation of missing keys to be controlled by the user.
By default set(o, key, value) will raise a KeyError if any of the key's
components are missing from the graphs of objects whereas unflatten(l)
will create a dict for missing keys.

Both functions can be passed a container factory that can be used to
create a different sort of key depending on key value, depth or some
schema only known to the application.